Add dummy 'call' module with the new tests
